챌린지 삭제 가능 조건 추가

+++ app-server/subprojects/bounded_context/challenge/application/src/main/kotlin/club/staircrusher/challenge/application/port/in/use_case/DeleteChallengeUseCase.kt
@@ -1,17 +1,26 @@
 package club.staircrusher.challenge.application.port.`in`.use_case
 
+import club.staircrusher.challenge.application.port.out.persistence.ChallengeContributionRepository
+import club.staircrusher.challenge.application.port.out.persistence.ChallengeParticipationRepository
 import club.staircrusher.challenge.application.port.out.persistence.ChallengeRepository
 import club.staircrusher.stdlib.di.annotation.Component
+import club.staircrusher.stdlib.domain.SccDomainException
 import club.staircrusher.stdlib.persistence.TransactionManager
 
 @Component
 class DeleteChallengeUseCase(
     private val transactionManager: TransactionManager,
     private val challengeRepository: ChallengeRepository,
+    private val challengeParticipationRepository: ChallengeParticipationRepository,
+    private val challengeContributionRepository: ChallengeContributionRepository,
 ) {
     fun handle(challengeId: String) = transactionManager.doInTransaction {
-        // TODO: 삭제 불가능 여부 체크하기; e.g. 이미 join한 유저가 있는 등.
-        // TODO: 필요한 경우 다른 리소스도 삭제하기; e.g. challenge condition?
+        if (challengeParticipationRepository.findByChallengeId(challengeId).isNotEmpty()) {
+            throw SccDomainException("이미 참여자가 존재하는 챌린지는 삭제할 수 없습니다.")
+        }
+        if (challengeContributionRepository.findByChallengeId(challengeId).isNotEmpty()) {
+            throw SccDomainException("이미 기여가 있는 챌린지는 삭제할 수 없습니다.")
+        }
         challengeRepository.remove(challengeId)
     }
 }
